factor out some common code into l1-parts
